Fix warnings caused by implicit type cast
warning C4267: '+=': conversion from 'size_t' to 'long', possible loss of data
warning C4244: '+=': conversion from '__int64' to 'i32', possible loss of data

+++ src/yyjson.c
@@ -5136,7 +5136,7 @@ yyjson_doc *yyjson_read_file(const char *path,
             }
             tmp = ((u8 *)buf) + buf_size - YYJSON_PADDING_SIZE - chunk_now;
             read_size = fread_safe(tmp, chunk_now, file);
-            file_size += read_size;
+            file_size += (long)read_size;
             if (read_size != chunk_now) break;
             
             chunk_now *= 2;
@@ -5717,7 +5717,7 @@ static_noinline u8 *write_f64_raw(u8 *buf, u64 raw, bool allow_nan_and_inf) {
             buf--;
             exp_dec++;
         } while (*buf == '0');
-        exp_dec += buf - hdr - 2;
+        exp_dec += (i32)(buf - hdr - 2);
         buf += (*buf != '.');
         buf[0] = 'e';
         buf++;
